Address more CRs

+++ IdentityCore/src/requests/MSIDSilentTokenRequest.m
@@ -188,7 +188,7 @@ typedef NS_ENUM(NSInteger, MSIDRefreshTokenTypes)
             if (accessToken.cachedAt)
             {
                 NSTimeInterval elapsed = [[NSDate date] timeIntervalSinceDate:accessToken.expiresOn];
-                [[MSIDExecutionFlowLogger sharedInstance] insertTag:@"1cd7n" extraInfo:@{@"d":@((int)elapsed)} withCorrelationId:self.requestParameters.correlationId];
+                [[MSIDExecutionFlowLogger sharedInstance] insertTag:@"1cd7n" extraInfo:@{@"d":@((int64_t)elapsed)} withCorrelationId:self.requestParameters.correlationId];
             }
         }
